# docs/swagger.yaml
definitions:
...
  handlers.CreateShareRequest:
    properties:
      expires_in_minutes:
        description: 以分钟为单位
        type: integer
      file_id:
        type: integer
      password:
        type: string
    required:
    - file_id
    type: object
...
  handlers.ShareCheckPasswordRequest:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  xerr.Response:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        description: 响应数据
      message:
        description: 消息
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/shares:
    post:
      consumes:
      - application/json
      description: 为指定文件或文件夹创建可分享链接，可设置密码和有效期
      parameters:
      - description: 分享链接信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateShareRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 分享链接创建成功
          schema:
            $ref: '#/definitions/xerr.Response'
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/xerr.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/xerr.Response'
        "403":
          description: 无权操作或文件状态异常
          schema:
            $ref: '#/definitions/xerr.Response'
        "409":
          description: 文件已存在有效分享链接
          schema:
            $ref: '#/definitions/xerr.Response'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/xerr.Response'
      security:
      - BearerAuth: []
      summary: 创建分享链接
      tags:
      - 分享
  /api/v1/shares/{share_id}:
    delete:
      description: 根据分享 ID 撤销用户创建的分享链接
      parameters:
      - description: 分享链接 ID
        in: path
        name: share_id
        required: true
        type: integer
      responses:
        "204":
          description: 分享链接撤销成功
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/xerr.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/xerr.Response'
        "403":
          description: 无权操作或链接已失效
          schema:
            $ref: '#/definitions/xerr.Response'
        "404":
          description: 分享链接不存在
          schema:
            $ref: '#/definitions/xerr.Response'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/xerr.Response'
      security:
      - BearerAuth: []
      summary: 撤销分享链接
      tags:
      - 分享
  /api/v1/shares/my:
    get:
      description: 列出当前用户创建的所有有效分享链接
      parameters:
      - default: 1
        description: 页码，默认为1
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量，默认为10
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 分享链接列表
          schema:
            properties:
              data:
                items:
                  $ref: '#/definitions/xerr.Response'
                type: array
              total:
                type: integer
            type: object
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/xerr.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/xerr.Response'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/xerr.Response'
      security:
      - BearerAuth: []
      summary: 列出用户创建的分享链接
      tags:
      - 分享
  /api/v1/user/me:
    get:
      consumes:
      - application/json
      description: 检索已认证用户的资料详情。
      produces:
      - application/json
      responses:
        "200":
          description: 用户资料检索成功
          schema:
            $ref: '#/definitions/xerr.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/xerr.Response'
        "404":
          description: 用户未找到
          schema:
            $ref: '#/definitions/xerr.Response'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/xerr.Response'
      security:
      - BearerAuth: []
      summary: 获取当前用户资料
      tags:
      - User
  /share/{share_uuid}/details:
    get:
      description: 根据分享 UUID 获取分享链接的详细信息（不包括文件内容），用于展示给下载者
      parameters:
      - description: 分享链接 UUID
        in: path
        name: share_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 分享链接详情
          schema:
            $ref: '#/definitions/xerr.Response'
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/xerr.Response'
        "403":
          description: 分享链接需要密码
          schema:
            $ref: '#/definitions/xerr.Response'
        "404":
          description: 分享链接不存在或已失效
          schema:
            $ref: '#/definitions/xerr.Response'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/xerr.Response'
      summary: 获取分享链接详情
      tags:
      - 分享
  /share/{share_uuid}/download:
    get:
      description: 根据分享 UUID 下载文件或文件夹（如果为文件夹则打包为 ZIP）
      parameters:
      - description: 分享链接 UUID
        in: path
        name: share_uuid
        required: true
        type: string
      - description: 分享密码（如果需要）
        in: query
        name: password
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 文件/文件夹下载成功
          schema:
            type: file
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/xerr.Response'
        "403":
          description: 分享链接需要密码或密码不正确
          schema:
            $ref: '#/definitions/xerr.Response'
        "404":
          description: 分享链接不存在或已失效
          schema:
            $ref: '#/definitions/xerr.Response'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/xerr.Response'
      summary: 下载分享内容
      tags:
      - 分享
  /share/{share_uuid}/verify:
    post:
      consumes:
      - application/json
      description: 验证分享链接的访问密码
      parameters:
      - description: 分享链接 UUID
        in: path
        name: share_uuid
        required: true
        type: string
      - description: 密码
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ShareCheckPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 密码验证成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 请求参数无效
          schema:
            $ref: '#/definitions/xerr.Response'
        "403":
          description: 密码不正确或链接已过期
          schema:
            $ref: '#/definitions/xerr.Response'
        "404":
          description: 分享链接不存在或已失效
          schema:
            $ref: '#/definitions/xerr.Response'
        "500":
          description: 内部服务器错误
          schema:
            $ref: '#/definitions/xerr.Response'
      summary: 验证分享链接密码
      tags:
      - 分享
swagger: "2.0"